requirements.txt
attrdict==2.0.0
certifi==2018.11.29
chardet==3.0.4
cytoolz==0.9.0.1
defusedxml==0.5.0
Django==2.1.4
django-allauth==0.38.0
django-extensions==2.1.4
django-rest-auth==0.9.3
djangorestframework==3.9.0
eth-abi==1.3.0
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-typing==2.0.0
eth-utils==1.4.1
hexbytes==0.1.0
idna==2.8
lru-dict==1.1.6
oauthlib==2.1.0
parsimonious==0.8.1
pycryptodome==3.7.2
python3-openid==3.1.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
rlp==1.0.3
six==1.12.0
toolz==0.9.0
urllib3==1.24.1
web3==4.8.2
websockets==6.0
